part of #5370  but would probably need more help from `vctrs` (probably https://github.com/r-lib/vctrs/issues/1169).