@romainfrancois 2020-07-01
This is a performance improvement. To match performance from 0.8.5 with list, we would need some sort of option in `vctrs::vec_rbind()` that would let it special case when an element of `...` is `vec_is_list()`.  

However in 1.0.0 we introduced some special handling that would "force" tibbles when we would give vectors or bare lists as part of `...`. This reverts it. 

I don't think it is a big deal but that is potentially a breaking change, or at least might create some friction with `expect_equal()`