@romainfrancois 2020-07-01
The special casing of bare lists is because this: 

``` r
str(vctrs::vec_rbind(
  list(a = 1, b = 2), 
  list(a = 3, b = 4)
))
#> 'data.frame':    2 obs. of  2 variables:
#>  $ a:List of 2
#>   ..$ : num 1
#>   ..$ : num 3
#>  $ b:List of 2
#>   ..$ : num 2
#>   ..$ : num 4

str(dplyr::bind_rows(
  list(a = 1, b = 2), 
  list(a = 3, b = 4)
))
#> 'data.frame':    2 obs. of  2 variables:
#>  $ a: num  1 3
#>  $ b: num  2 4
```

<sup>Created on 2020-07-01 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9001)</sup>

but this is something for `vctrs` to address, presumably here: https://github.com/r-lib/vctrs/issues/1169