@romainfrancois 2020-07-01
With this code: 

```r
library(tibble)
library(bench)
f <- function(){
  library(dplyr, warn.conflicts = FALSE)
  library(purrr)
  
  new_list <- map(1:10000, function(x) {
    list(A = 2 * x, B = 4 * x)
  })
  
  bench::mark(bind_rows(new_list))
}
callr::r(f)
```

we get these results with this branch: 

```
[bind_rows_list_perf] 205.2 MiB ❯ callr::r(f)
# A tibble: 1 x 13
  expression               min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result                memory                     time         gc              
  <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>                <list>                     <list>       <list>          
1 bind_rows(new_list)    173ms    228ms      4.64       1MB     35.5     3    23      647ms <df[,2] [10,000 × 2]> <Rprofmem[,3] [1,049 × 3]> <bch:tm [3]> <tibble [3 × 3]>
```

these with master: 

```
[master] 164.3 MiB ❯ callr::r(f)
# A tibble: 1 x 13
  expression               min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result                memory                      time         gc              
  <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>                <list>                      <list>       <list>          
1 bind_rows(new_list)    5.53s    5.53s     0.181    9.38MB     19.7     1   109      5.53s <tibble [10,000 × 2]> <Rprofmem[,3] [34,092 × 3]> <bch:tm [1]> <tibble [1 × 3]>
```

it isn't quite the performance of dplyr 0.8.5: 

```
[master] 196.4 MiB ❯ callr::r(f, libpath = "/Users/romainfrancois/git/tidyverse/bench-libs/0.8.5/")
# A tibble: 1 x 13
  expression               min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result                memory time          gc               
  <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>                <list> <list>        <list>           
1 bind_rows(new_list)   6.68ms   7.59ms      134.        NA     143.    28    30      210ms <tibble [10,000 × 2]> <NULL> <bch:tm [58]> <tibble [58 × 3]>
```

but to get there `dplyr` needs help from `vctrs` , as most of the loss is due to converting the lists to data frames so that `vec_rbind()` makes vectors instead of lists.