@romainfrancois 2020-07-01
The overall file looks suspicious, but the new problems section gives: 

https://github.com/tidyverse/dplyr/blob/27100225172b9391ed8ea7942bf97a4c39390528/revdep/README.md#new-problems-36

```
|package                                              |version |error  |warning |note |
|:----------------------------------------------------|:-------|:------|:-------|:----|
|[amt](problems.md#amt)                               |0.1.2   |__+1__ |        |     |
|[anglr](problems.md#anglr)                           |0.6.0   |__+1__ |        |     |
|[baguette](problems.md#baguette)                     |0.0.1   |__+1__ |        |     |
|[c14bazAAR](problems.md#c14bazaar)                   |1.2.0   |__+1__ |        |     |
|[coalitions](problems.md#coalitions)                 |0.6.12  |__+1__ |        |     |
|[crimedata](problems.md#crimedata)                   |0.2.0   |__+1__ |        |     |
|[dbparser](problems.md#dbparser)                     |1.1.2   |__+1__ |        |     |
|[DiversityOccupancy](problems.md#diversityoccupancy) |1.0.6   |       |__+1__  |     |
|[easyalluvial](problems.md#easyalluvial)             |0.2.3   |__+1__ |        |     |
|[eph](problems.md#eph)                               |0.4.0   |__+1__ |        |2    |
|[evaluator](problems.md#evaluator)                   |0.4.2   |__+1__ |        |     |
|[fabletools](problems.md#fabletools)                 |0.2.0   |__+1__ |        |1    |
|[fauxnaif](problems.md#fauxnaif)                     |0.5.6   |__+1__ |        |     |
|[forestmangr](problems.md#forestmangr)               |0.9.2   |__+1__ |        |1    |
|[ftExtra](problems.md#ftextra)                       |0.0.1   |__+1__ |        |     |
|[ggpubr](problems.md#ggpubr)                         |0.4.0   |__+1__ |        |     |
|[gravity](problems.md#gravity)                       |0.9.8   |__+2__ |        |     |
|[keyATM](problems.md#keyatm)                         |0.2.2   |__+1__ |        |2    |
|[kntnr](problems.md#kntnr)                           |0.4.4   |__+1__ |        |     |
|[nanny](problems.md#nanny)                           |0.1.8   |__+2__ |        |1    |
|[pmdplyr](problems.md#pmdplyr)                       |0.3.3   |__+1__ |        |     |
|[psychmeta](problems.md#psychmeta)                   |2.3.10  |__+1__ |        |     |
|[rnoaa](problems.md#rnoaa)                           |1.0.0   |__+1__ |        |     |
|[SEERaBomb](problems.md#seerabomb)                   |2019.2  |       |__+1__  |     |
|[sigminer](problems.md#sigminer)                     |1.0.7   |__+1__ |        |     |
|[simTool](problems.md#simtool)                       |1.1.6   |__+1__ |        |     |
|[survminer](problems.md#survminer)                   |0.4.7   |__+1__ |        |1    |
|[SwimmeR](problems.md#swimmer)                       |0.2.0   |__+1__ |        |1    |
|[tibbletime](problems.md#tibbletime)                 |0.1.5   |__+1__ |        |     |
|[tidyboot](problems.md#tidyboot)                     |0.1.1   |__+1__ |        |     |
|[timetk](problems.md#timetk)                         |2.0.0   |__+1__ |        |2    |
|[unheadr](problems.md#unheadr)                       |0.2.1   |__+1__ |        |     |
|[unpivotr](problems.md#unpivotr)                     |0.6.0   |__+1__ |        |     |
|[varsExplore](problems.md#varsexplore)               |0.1.0   |__+1__ |        |1    |
|[viafr](problems.md#viafr)                           |0.2.0   |__+1__ |        |1    |
|[xpose](problems.md#xpose)                           |0.4.10  |__+2__ |        |     |
```